+++ NCU-RoboMaster/yaw_fdb_encoder/MDK-ARM/Communication/minipc.h
@@ -7,10 +7,10 @@
 typedef struct{
 
 unsigned char 		frame_header; 		  //帧头0xFF
-	float 					angle_yaw;     			//yaw angle
-	float 					angle_pit;     			//pitch angle 
+ int16_t 					angle_yaw;     			//yaw angle
+ int16_t 					angle_pit;     			//pitch angle 
 unsigned char 		state_flag;     		//当前状态：【0 未瞄准目标 】【 1 近距离锁定目标】【2 远距离锁定目标】【3 打符模式】
-	float 					distance;     			//目标距离
+	int16_t 					distance;     			//目标距离
 unsigned char 		frame_tail; 	  	  //帧尾0xFE
 }Minipc_Rx;
